requirements.txt
streamlit
tensorflow-cpu
numpy
pillow
scikit-learn